requirements.txt
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
crewai>=0.28.0
